modificar text_es.js.y reacomodar la arquitectura
