+++ src/metodos_numericos_dcb_fi/utilidades/__init__.py
@@ -1,2 +1,8 @@
-from metodos_numericos_dcb_fi.utilidades.util import *
+# ------------------- Importar módulo de configuracion -------------------
+from metodos_numericos_dcb_fi.utilidades.configuracion import *
 leerTextos()
+# ------------------- Importar el resto de módulos -------------------
+from metodos_numericos_dcb_fi.utilidades.extras import *
+from metodos_numericos_dcb_fi.utilidades.leerEntrada import *
+from metodos_numericos_dcb_fi.utilidades.mostrarResultados import *
+from metodos_numericos_dcb_fi.utilidades.validacion import *
